requirements.txt
# 可选依赖：仅 oc_ast_splitter 需要
libclang>=14.0.0
